added fixed RSI indicator and grouped update for indicators

+++ indicators/SMA/dsma.py
@@ -5,6 +5,6 @@ def dsma(sma, data, index):
 		return(None)
 	if (sma == 0 or data[index][0] == 0):
 		return(None)
-	d = float(Decimal(data[index][0])/Decimal(sma))
+	d = Decimal(data[index][0])/Decimal(sma)
 	result = (d - 1) * 100
-	return(result)
+	return(float(result))
