+++ indicators/VolumeMA/VolMA_calculator.py
@@ -20,4 +20,4 @@ def VolMA_calculator(period, data, index):
 		else:
 			tmp = tmp + data[index-i][0]
 		i+=1
-	return(float(Decimal(tmp)/Decimal(period)))
+	return(Decimal(tmp)/Decimal(period))
